Avoid extra string decoding in Parquet writer
This patch is supposed to resolve performance issues for the Parquet
writer.

Strings are provided in as `Text` that contains byte array with
UTF-8 bytes internally. Instead of decoding those bytes to `String`
and encoding them we can pass the byte array taken from Text
directly to the record consumer.